TensorYlm: move symmetry computation to helpers
