+++ src/NumericalAlgorithms/SphericalHarmonics/TensorYlmHelpers.cpp
@@ -7,6 +7,7 @@
 #include <complex>
 #include <limits>
 
+#include "DataStructures/Tensor/TypeAliases.hpp"
 #include "Utilities/Array.hpp"
 #include "Utilities/ErrorHandling/Assert.hpp"
 #include "Utilities/Gsl.hpp"
@@ -71,6 +72,29 @@ std::complex<double> bv_to_k(const BasisVector basis_vector, const int i) {
   return result;
 }
 
+template <typename Symm>
+double get_symm_factor(const size_t src_multiplicity, const size_t lbar) {
+  static_assert(std::is_same_v<Symmetry<3, 2, 1>, Symm> or
+                    std::is_same_v<Symmetry<2, 1, 1>, Symm> or
+                    std::is_same_v<Symmetry<2, 1>, Symm> or
+                    std::is_same_v<Symmetry<1, 1>, Symm>,
+                "Unimplemented symmetry");
+  if constexpr (std::is_same_v<Symmetry<3, 2, 1>, Symm> or
+                std::is_same_v<Symmetry<2, 1>, Symm>) {
+    // Not symmetric on the 2 indices.
+    (void)src_multiplicity;
+    (void)lbar;
+  } else {
+    // symmetric on the 2 indices.
+    //
+    // Note that src_multiplicity is 1 if the indices are the same,
+    // 2 if they are different.  For more complicated symmetries, if
+    // we ever wanted to implement them, this formula would be different.
+    return (lbar % 2 == 0 ? static_cast<double>(src_multiplicity) : 0.0);
+  }
+  return 1.0;
+}
+
 // Explicit instantiations
 template std::array<BasisVector, 1> to_cart_basis_vector(
     const cpp20::array<size_t, 1>& indices);
@@ -78,5 +102,13 @@ template std::array<BasisVector, 2> to_cart_basis_vector(
     const cpp20::array<size_t, 2>& indices);
 template std::array<BasisVector, 3> to_cart_basis_vector(
     const cpp20::array<size_t, 3>& indices);
+template double get_symm_factor<Symmetry<3, 2, 1>>(size_t src_multiplicity,
+                                                   size_t lbar);
+template double get_symm_factor<Symmetry<2, 1, 1>>(size_t src_multiplicity,
+                                                   size_t lbar);
+template double get_symm_factor<Symmetry<1, 1>>(size_t src_multiplicity,
+                                                size_t lbar);
+template double get_symm_factor<Symmetry<2, 1>>(size_t src_multiplicity,
+                                                size_t lbar);
 
 }  // namespace ylm::TensorYlm::helpers
