+++ src/NumericalAlgorithms/SphericalHarmonics/TensorYlmHelpers.hpp
@@ -27,4 +27,8 @@ int bv_to_m(BasisVector basis_vector, int i);
 /// Returns the prefactor k associated with a Cartesian basis vector.
 std::complex<double> bv_to_k(BasisVector basis_vector, int i);
 
+/// Computes the symmetry factor S that appears in various equations.
+template <typename Symm>
+double get_symm_factor(size_t src_multiplicity, size_t lbar);
+
 }  // namespace ylm::TensorYlm::helpers
